GGB-641: implement semi-numerical computations (currently only conics are supported)
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@49768 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/kernel/prover/polynomial/Polynomial.java
@@ -63,7 +63,7 @@ public class Polynomial implements Comparable<Polynomial> {
 	 * @param coeff
 	 *            the constant
 	 */
-	public Polynomial(final int coeff) {
+	public Polynomial(final long coeff) {
 		this(coeff, new Term());
 	}
